requirements.txt
asgiref==3.2.3
blis==0.2.4
certifi==2019.11.28
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
cymem==2.0.3
defusedxml==0.6.0
Django==3.0.2
django-appconf==1.0.3
django-crispy-forms==1.8.1
django-multiselectfield==0.1.11
django-select2==7.2.0
idna==2.8
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
Pillow==7.0.0
Pint==0.11
pkg-resources==0.0.0
plac==0.9.6
preshed==2.0.1
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2019.3
PyYAML==3.13
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.3.0
srsly==1.0.1
thinc==7.0.8
tqdm==4.43.0
urllib3==1.25.8
wasabi==0.6.0
